clone norm
