--- src/main/java/com/dmitriyevseyev/car-manager2/exceptions/AddCarExeption.java
@@ -1,11 +1,0 @@
-package exceptions;
-
-public class AddCarExeption extends Exception {
-    public AddCarExeption() {
-        super();
-    }
-
-    public AddCarExeption(String message) {
-        super(message);
-    }
-}
