--- src/main/java/com/dmitriyevseyev/car-manager2/exceptions/DeleteCarExeption.java
@@ -1,12 +1,0 @@
-package exceptions;
-
-public class DeleteCarExeption extends Exception{
-
-    public DeleteCarExeption() {
-        super();
-    }
-
-    public DeleteCarExeption(String message) {
-        super(message);
-    }
-}
