--- src/main/java/com/dmitriyevseyev/car-manager2/exceptions/GetAllCarExeption.java
@@ -1,11 +1,0 @@
-package exceptions;
-
-public class GetAllCarExeption extends Exception {
-    public GetAllCarExeption() {
-        super();
-    }
-
-    public GetAllCarExeption(String message) {
-        super(message);
-    }
-}
