--- src/main/java/com/dmitriyevseyev/car-manager2/exceptions/NotFoundException.java
@@ -1,12 +1,0 @@
-package exceptions;
-
-public class NotFoundException extends Exception {
-
-    public NotFoundException() {
-        super();
-    }
-
-    public NotFoundException(String message) {
-        super(message);
-    }
-}
