--- src/main/java/com/dmitriyevseyev/car-manager2/exceptions/UpdateCarException.java
@@ -1,12 +1,0 @@
-package exceptions;
-
-public class UpdateCarException extends Exception {
-
-    public UpdateCarException() {
-        super();
-    }
-
-    public UpdateCarException (String message) {
-        super(message);
-    }
-}
