--- src/main/java/com/dmitriyevseyev/car-manager2/view/carEdit.fxml
@@ -1,61 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.text.*?>
-
-<AnchorPane prefHeight="316.0" prefWidth="495.0" xmlns="http://javafx.com/javafx/16" xmlns:fx="http://javafx.com/fxml/1" fx:controller="view.EditCarController">
-   <children>
-      <GridPane layoutX="19.0" layoutY="10.0" prefHeight="250.0" prefWidth="455.0">
-         <columnConstraints>
-            <ColumnConstraints hgrow="SOMETIMES" maxWidth="289.0" minWidth="10.0" prefWidth="173.0" />
-            <ColumnConstraints hgrow="SOMETIMES" maxWidth="408.0" minWidth="10.0" prefWidth="331.0" />
-         </columnConstraints>
-         <rowConstraints>
-            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-         </rowConstraints>
-         <children>
-            <Label alignment="TOP_LEFT" prefHeight="36.0" prefWidth="185.0" text="Id" textAlignment="CENTER">
-               <font>
-                  <Font size="25.0" />
-               </font>
-            </Label>
-            <Label alignment="TOP_LEFT" prefHeight="36.0" prefWidth="176.0" text="Name" GridPane.rowIndex="1">
-               <font>
-                  <Font size="25.0" />
-               </font>
-            </Label>
-            <Label alignment="TOP_LEFT" prefHeight="24.0" prefWidth="173.0" text="Color" textAlignment="CENTER" GridPane.rowIndex="3">
-               <font>
-                  <Font size="25.0" />
-               </font>
-            </Label>
-            <Label alignment="TOP_LEFT" prefHeight="36.0" prefWidth="193.0" text="isAfterCrash" GridPane.rowIndex="4">
-               <font>
-                  <Font size="25.0" />
-               </font>
-            </Label>
-            <TextField fx:id="idField" GridPane.columnIndex="1" />
-            <TextField fx:id="nameField" GridPane.columnIndex="1" GridPane.rowIndex="1" />
-            <TextField fx:id="colorField" GridPane.columnIndex="1" GridPane.rowIndex="3" />
-            <CheckBox fx:id="isAfterCrashField" mnemonicParsing="false" GridPane.columnIndex="1" GridPane.rowIndex="4" />
-            <DatePicker fx:id="dp" editable="false" onAction="#ShowDate" prefHeight="25.0" prefWidth="308.0" GridPane.columnIndex="1" GridPane.rowIndex="2" />
-            <Label text="Date" GridPane.rowIndex="2">
-               <font>
-                  <Font size="25.0" />
-               </font>
-            </Label>
-         </children>
-      </GridPane>
-      <ButtonBar layoutX="231.0" layoutY="260.0" prefHeight="40.0" prefWidth="234.0" AnchorPane.bottomAnchor="16.0" AnchorPane.leftAnchor="231.0" AnchorPane.rightAnchor="30.0" AnchorPane.topAnchor="260.0">
-        <buttons>
-            <Button mnemonicParsing="false" onAction="#handleOk" prefHeight="25.0" text="OK" />
-          <Button alignment="TOP_CENTER" contentDisplay="CENTER" mnemonicParsing="false" onAction="#handleCancel" prefHeight="25.0" prefWidth="120.0" text="Cancel" />
-        </buttons>
-      </ButtonBar>
-   </children>
-</AnchorPane>
