--- src/main/java/com/dmitriyevseyev/car-manager2/view/view.fxml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.text.*?>
-
-<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="472.0" prefWidth="617.0" xmlns="http://javafx.com/javafx/16" xmlns:fx="http://javafx.com/fxml/1" fx:controller="view.ControllerView">
-    <children>
-        <Button fx:id="add" layoutX="103.0" layoutY="409.0" mnemonicParsing="false" onAction="#addRow" prefHeight="25.0" prefWidth="99.0" text="Создать" />
-        <Button fx:id="deleteCar" disable="true" layoutX="415.0" layoutY="409.0" mnemonicParsing="false" onAction="#deleteSelectedRows" prefHeight="25.0" prefWidth="99.0" text="Удалить" />
-        <Button fx:id="modifyCar" disable="true" layoutX="260.0" layoutY="409.0" mnemonicParsing="false" onAction="#editSelectedRow" prefHeight="25.0" prefWidth="99.0" text="Редактировать" />
-        <TableView fx:id="tableview" accessibleRole="IMAGE_VIEW" editable="true" layoutX="43.0" layoutY="81.0" prefHeight="288.0" prefWidth="532.0">
-            <columns>
-                <TableColumn fx:id="actionColumn" prefWidth="44.0" text="action" />
-                <TableColumn fx:id="idColumn" prefWidth="59.0" text="id" />
-                <TableColumn fx:id="nameColumn" prefWidth="143.0" text="name" />
-                <TableColumn fx:id="dateColumn" prefWidth="93.0" text="date" />
-                <TableColumn fx:id="colorColumn" prefWidth="101.0" text="color" />
-                <TableColumn fx:id="isAfterCrashColumn" prefWidth="91.0" text="isAfterCrash" />
-            </columns>
-         <columnResizePolicy>
-            <TableView fx:constant="CONSTRAINED_RESIZE_POLICY" />
-         </columnResizePolicy>
-        </TableView>
-        <Label fx:id="carManager" alignment="CENTER" layoutX="228.0" layoutY="28.0" prefHeight="17.0" prefWidth="162.0" text="CAR MANAGER" textAlignment="CENTER">
-            <font>
-                <Font size="22.0" />
-            </font>
-        </Label>
-    </children>
-</Pane>
