+++ src/main/java/com/dmitriyevseyev/carmanager2/exceptions/AddCarExeption.java
@@ -1,0 +1,11 @@
+package com.dmitriyevseyev.carmanager2.exceptions;
+
+public class AddCarExeption extends Exception {
+    public AddCarExeption() {
+        super();
+    }
+
+    public AddCarExeption(String message) {
+        super(message);
+    }
+}
