+++ src/main/java/com/dmitriyevseyev/carmanager2/exceptions/DeleteCarExeption.java
@@ -1,0 +1,12 @@
+package com.dmitriyevseyev.carmanager2.exceptions;
+
+public class DeleteCarExeption extends Exception{
+
+    public DeleteCarExeption() {
+        super();
+    }
+
+    public DeleteCarExeption(String message) {
+        super(message);
+    }
+}
