+++ src/main/java/com/dmitriyevseyev/carmanager2/exceptions/GetAllCarExeption.java
@@ -1,0 +1,11 @@
+package com.dmitriyevseyev.carmanager2.exceptions;
+
+public class GetAllCarExeption extends Exception {
+    public GetAllCarExeption() {
+        super();
+    }
+
+    public GetAllCarExeption(String message) {
+        super(message);
+    }
+}
