+++ src/main/java/com/dmitriyevseyev/carmanager2/exceptions/NotFoundException.java
@@ -1,0 +1,12 @@
+package com.dmitriyevseyev.carmanager2.exceptions;
+
+public class NotFoundException extends Exception {
+
+    public NotFoundException() {
+        super();
+    }
+
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
