+++ src/main/java/com/dmitriyevseyev/carmanager2/exceptions/UpdateCarException.java
@@ -1,0 +1,12 @@
+package com.dmitriyevseyev.carmanager2.exceptions;
+
+public class UpdateCarException extends Exception {
+
+    public UpdateCarException() {
+        super();
+    }
+
+    public UpdateCarException (String message) {
+        super(message);
+    }
+}
