+++ src/main/java/com/dmitriyevseyev/carmanager2/model/Car.java
@@ -1,0 +1,150 @@
+package com.dmitriyevseyev.carmanager2.model;
+
+
+import java.util.Date;
+import java.util.Objects;
+
+public class Car {
+    private Integer id;
+    private String name;
+    private Date date;
+    private String color;
+    private boolean isAfterCrash;
+
+    public Car() {
+    }
+
+    public Car(Integer id, String name, Date date, String color, boolean isAfterCrash) {
+        this.id = id;
+        this.name = name;
+        this.date = date;
+        this.color = color;
+        this.isAfterCrash = isAfterCrash;
+    }
+
+    public Integer getId() {
+        return id;
+    }
+
+    public void setId(Integer id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public Date getDate() {
+        return date;
+    }
+
+    public void setDate(Date date) {
+        this.date = date;
+    }
+
+    public String getColor() {
+        return color;
+    }
+
+    public void setColor(String color) {
+        this.color = color;
+    }
+
+    public boolean isAfterCrash() {
+        return isAfterCrash;
+    }
+
+    public void setAfterCrash(boolean afterCrash) {
+        isAfterCrash = afterCrash;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Car car = (Car) o;
+        return isAfterCrash == car.isAfterCrash &&
+                Objects.equals(id, car.id) &&
+                Objects.equals(name, car.name) &&
+                Objects.equals(date, car.date) &&
+                Objects.equals(color, car.color);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id, name, date, color, isAfterCrash);
+    }
+
+    @Override
+    public String toString() {
+        return "Car{" +
+                "id=" + id +
+                ", name='" + name + '\'' +
+                ", date=" + date +
+                ", color='" + color + '\'' +
+                ", isAfterCrash=" + isAfterCrash +
+                '}';
+    }
+
+    // builder pattern read more about it
+    // reference: https://projectlombok.org/features/Builder
+    public static Builder builder() {
+        return new Builder();
+    }
+
+    public static class Builder {
+        private Integer id;
+        private String name;
+        private Date date;
+        private String color;
+        private boolean isAfterCrash;
+
+        Builder() {
+        }
+
+        public Builder id(Integer id) {
+            this.id = id;
+
+            // why we return this? it's chain pattern. you can read more about it
+            return this;
+        }
+
+        public Builder name(String name) {
+            this.name = name;
+
+            return this;
+        }
+
+        public Builder date(Date date) {
+            this.date = date;
+
+            return this;
+        }
+
+        public Builder color(String color) {
+            this.color = color;
+
+            return this;
+        }
+
+        public Builder isAfterCrash(boolean isAfterCrash) {
+            this.isAfterCrash = isAfterCrash;
+
+            return this;
+        }
+
+        public Car build() {
+            return new Car(
+                    id,
+                    name,
+                    date,
+                    color,
+                    isAfterCrash
+            );
+        }
+    }
+}
